refactor(web): extract shared rating picker popover to eliminate duplication
