+++ apps/web/src/entities/book/ui/BookCoverCard/BookCoverCard.module.scss
@@ -303,27 +303,3 @@
   font-size: 13px;
 }
 
-// ─── Поп-овер выбора оценки ───────────────────────────────────────────────────
-
-.cover-card__rating-picker {
-  width: 200px;
-  padding: 16px 16px 4px;
-}
-
-.cover-card__rating-stars {
-  display: flex;
-  justify-content: center;
-  margin-bottom: 8px;
-}
-
-.cover-card__rating-value {
-  text-align: center;
-  font-weight: 600;
-  font-size: 14px;
-  margin-bottom: 4px;
-}
-
-.cover-card__rating-value-denom {
-  font-size: 12px;
-  opacity: 0.55;
-}
